Create breach_api.py
sign up on rapid API for a key, the error checking in this is ABSENT !! but it it setup to accept an email, the query option is predefined but easily changeable with the addition of an entry and set that on line 23 replacing auto. and drop the data. have not setup a *NIX path yet, windows at the moment.